refactor: improve logging and organizade structure by usecases
