+++ docs/features/LeetCodeDailyTask.md
@@ -2,13 +2,13 @@
 
 [<< README.md](../../README.md)
 
-This feature automatically fetches and posts the LeetCode Problem of the Day to a specified Discord channel every day at 3:00 AM.
+This feature automatically fetches and posts the LeetCode Problem of the Day to a specified Discord channel every day at 15:00 UTC.
 
 ## Features
 
-- **Automatic Daily Posts**: Sends the LeetCode problem of the day at 3:00 AM daily
+- **Automatic Daily Posts**: Sends the LeetCode problem of the day at 15:00 UTC daily
 - **Rich Embed Messages**: Beautiful Discord embeds with problem details
-- **Manual Trigger**: Use `/leetcode` command to manually fetch today's problem
+- **Manual Trigger**: Use `/leetcode-daily` command to manually fetch today's problem
 - **Comprehensive Information**: Includes difficulty, acceptance rate, topics, and direct link
 
 ## Setup
@@ -27,13 +27,6 @@ To get your channel ID:
 3. Click "Copy ID"
 4. Paste this ID as the value for `LEETCODE_CHANNEL_ID`
 
-### 2. Install Dependencies
-
-The bot requires the `aiohttp` library for making HTTP requests to LeetCode's API:
-
-```bash
-pip install -r requirements.txt
-```
 
 ### 3. Bot Permissions
 
@@ -46,13 +39,13 @@ Ensure your bot has the following permissions in the target channel:
 
 ### Automatic Daily Posts
 
-The bot will automatically post the LeetCode problem of the day at 3:00 AM (server time) to the configured channel.
+The bot will automatically post the LeetCode problem of the day at 15:00 UTC to the configured channel.
 
 ### Manual Command
 
 Users can manually trigger the LeetCode problem fetch using:
 ```
-/leetcode
+/leetcode-daily
 ```
 
 This command can be used in any channel where the bot has permissions.
@@ -67,25 +60,6 @@ The bot sends an embed message containing:
 - **Topics**: Related algorithmic topics (up to 5)
 - **Direct Link**: Click to solve the problem on LeetCode
 
-## Troubleshooting
-
-### Common Issues
-
-1. **Channel ID not configured**
-   - Error: "LEETCODE_CHANNEL_ID not configured"
-   - Solution: Add the correct channel ID to your `.env` file
-
-2. **Channel not found**
-   - Error: "Channel not found: [ID]"
-   - Solution: Verify the channel ID is correct and the bot has access
-
-3. **Failed to fetch problem**
-   - Error: "Failed to fetch daily problem"
-   - Solution: Check internet connection and LeetCode availability
-
-### Debug Output
-
-The bot logs all LeetCode task activities with the prefix `[BOT][TASK][LEETCODE]`. Check your console output for detailed error information.
 
 ## Technical Details
 
@@ -94,17 +68,3 @@ The bot logs all LeetCode task activities with the prefix `[BOT][TASK][LEETCODE]
 The bot uses LeetCode's GraphQL API endpoint:
 - URL: `https://leetcode.com/graphql`
 - Query: Fetches the active daily coding challenge
-
-### Schedule
-
-The task uses Discord.py's `@tasks.loop()` decorator with `time(3, 0)` to run at 3:00 AM daily. The time is based on the server's timezone.
-
-### Error Handling
-
-The bot includes comprehensive error handling:
-- Network request failures
-- JSON parsing errors
-- Discord API errors
-- Missing configuration errors
-
-All errors are logged to the console with descriptive messages.
